requirements.txt
numpy
pandas
yfinance
joblib
requests
scikit-learn
tensorflow-cpu
h5py